Cleanup incorrect merge
