+++ src/main/java/de/brettspielwelt/game/Informer2.java
