+++ src/main/java/de/brettspielwelt/game/cabanga/Informer.java
