+++ src/main/java/de/brettspielwelt/game/hit/Informer.java
